# docs/api/openapi_v1.json
{
  "components": {
    "schemas": {
...
      "CashFlowAnalyticResponse": {
        "properties": {
          "as_of_date": {
            "format": "date",
            "title": "As Of Date",
            "type": "string"
          },
          "cashflow_items": {
            "items": {
              "$ref": "#/components/schemas/CashFlowItem"
            },
            "title": "Cashflow Items",
            "type": "array"
          },
          "total_net_cashflow": {
            "title": "Total Net Cashflow",
            "type": "string"
          }
        },
        "required": [
          "as_of_date",
          "cashflow_items",
          "total_net_cashflow"
        ],
        "title": "CashFlowAnalyticResponse",
        "type": "object"
      },
      "CashFlowBaselineSnapshotCreate": {
        "properties": {
          "as_of_date": {
            "format": "date",
            "title": "As Of Date",
            "type": "string"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          }
        },
        "required": [
          "as_of_date",
          "correlation_id"
        ],
        "title": "CashFlowBaselineSnapshotCreate",
        "type": "object"
      },
      "CashFlowBaselineSnapshotResponse": {
        "properties": {
          "as_of_date": {
            "format": "date",
            "title": "As Of Date",
            "type": "string"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "snapshot_data": {
            "title": "Snapshot Data",
            "type": "object"
          },
          "total_net_cashflow": {
            "title": "Total Net Cashflow",
            "type": "string"
          }
        },
        "required": [
          "id",
          "as_of_date",
          "snapshot_data",
          "total_net_cashflow",
          "created_at",
          "correlation_id"
        ],
        "title": "CashFlowBaselineSnapshotResponse",
        "type": "object"
      },
      "CashFlowCreate": {
        "properties": {
          "amount": {
            "description": "Derived cashflow amount",
            "title": "Amount",
            "type": "number"
          },
          "contract_id": {
            "description": "Source contract ID",
            "format": "uuid",
            "title": "Contract Id",
            "type": "string"
          },
          "value_date": {
            "description": "Cashflow value date",
            "format": "date",
            "title": "Value Date",
            "type": "string"
          },
          "view": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CashFlowView"
              }
            ],
            "description": "CashFlow view type"
          }
        },
        "required": [
          "contract_id",
          "view",
          "amount",
          "value_date"
        ],
        "title": "CashFlowCreate",
        "type": "object"
      },
      "CashFlowItem": {
        "properties": {
          "amount_usd": {
            "title": "Amount Usd",
            "type": "string"
          },
          "mtm_value": {
            "title": "Mtm Value",
            "type": "string"
          },
          "object_id": {
            "title": "Object Id",
            "type": "string"
          },
          "object_type": {
            "title": "Object Type",
            "type": "string"
          },
          "settlement_date": {
            "format": "date",
            "title": "Settlement Date",
            "type": "string"
          }
        },
        "required": [
          "object_type",
          "object_id",
          "settlement_date",
          "amount_usd",
          "mtm_value"
        ],
        "title": "CashFlowItem",
        "type": "object"
      },
      "CashFlowLedgerEntryRead": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "string"
          },
          "cashflow_date": {
            "format": "date",
            "title": "Cashflow Date",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "direction": {
            "title": "Direction",
            "type": "string"
          },
          "hedge_contract_id": {
            "format": "uuid",
            "title": "Hedge Contract Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "leg_id": {
            "title": "Leg Id",
            "type": "string"
          },
          "source_event_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Event Id"
          },
          "source_event_type": {
            "title": "Source Event Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "hedge_contract_id",
          "source_event_type",
          "source_event_id",
          "leg_id",
          "cashflow_date",
          "currency",
          "direction",
          "amount",
          "created_at"
        ],
        "title": "CashFlowLedgerEntryRead",
        "type": "object"
      },
      "CashFlowRead": {
        "properties": {
          "amount": {
            "description": "Derived cashflow amount",
            "title": "Amount",
            "type": "number"
          },
          "contract_id": {
            "description": "Source contract ID",
            "format": "uuid",
            "title": "Contract Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "value_date": {
            "description": "Cashflow value date",
            "format": "date",
            "title": "Value Date",
            "type": "string"
          },
          "view": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CashFlowView"
              }
            ],
            "description": "CashFlow view type"
          }
        },
        "required": [
          "contract_id",
          "view",
          "amount",
          "value_date",
          "id",
          "created_at"
        ],
        "title": "CashFlowRead",
        "type": "object"
      },
      "CashFlowView": {
        "enum": [
          "analytic",
          "baseline",
          "ledger",
          "what_if"
        ],
        "title": "CashFlowView",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HedgeContractCreate": {
        "properties": {
          "commodity": {
            "description": "Commodity identifier",
            "title": "Commodity",
            "type": "string"
          },
          "legs": {
            "description": "Exactly two legs: one fixed, one variable",
            "items": {
              "$ref": "#/components/schemas/HedgeLeg"
            },
            "title": "Legs",
            "type": "array"
          },
          "quantity_mt": {
            "description": "Quantity in metric tons (MT)",
            "title": "Quantity Mt",
            "type": "number"
          }
        },
        "required": [
          "commodity",
          "quantity_mt",
          "legs"
        ],
        "title": "HedgeContractCreate",
        "type": "object"
      },
      "HedgeContractRead": {
        "properties": {
          "classification": {
            "allOf": [
              {
                "$ref": "#/components/schemas/HedgeClassification"
              }
            ],
            "description": "Classification based on fixed leg"
          },
          "commodity": {
            "title": "Commodity",
            "type": "string"
          },
          "counterparty_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counterparty Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "fixed_leg_side": {
            "allOf": [
              {
                "$ref": "#/components/schemas/HedgeLegSide"
              }
            ],
            "description": "Fixed leg side (buy or sell)"
          },
          "fixed_price_unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fixed Price Unit"
          },
          "fixed_price_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fixed Price Value"
          },
          "float_pricing_convention": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Float Pricing Convention"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "quantity_mt": {
            "title": "Quantity Mt",
            "type": "number"
          },
          "rfq_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rfq Id"
          },
          "rfq_quote_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rfq Quote Id"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "variable_leg_side": {
            "allOf": [
              {
                "$ref": "#/components/schemas/HedgeLegSide"
              }
            ],
            "description": "Variable leg side (buy or sell)"
          }
        },
        "required": [
          "id",
          "commodity",
          "quantity_mt",
          "fixed_leg_side",
          "variable_leg_side",
          "classification",
          "created_at"
        ],
        "title": "HedgeContractRead",
        "type": "object"
      },
      "HedgeContractSettlementCreate": {
        "properties": {
          "cashflow_date": {
            "format": "date",
            "title": "Cashflow Date",
            "type": "string"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "legs": {
            "items": {
              "$ref": "#/components/schemas/HedgeContractSettlementLeg"
            },
            "title": "Legs",
            "type": "array"
          },
          "source_event_id": {
            "format": "uuid",
            "title": "Source Event Id",
            "type": "string"
          }
        },
        "required": [
          "source_event_id",
          "cashflow_date",
          "legs"
        ],
        "title": "HedgeContractSettlementCreate",
        "type": "object"
      },
...
      "HedgeContractSettlementResponse": {
        "properties": {
          "event": {
            "$ref": "#/components/schemas/HedgeContractSettlementEventRead"
          },
          "ledger_entries": {
            "items": {
              "$ref": "#/components/schemas/CashFlowLedgerEntryRead"
            },
            "title": "Ledger Entries",
            "type": "array"
          }
        },
        "required": [
          "event",
          "ledger_entries"
        ],
        "title": "HedgeContractSettlementResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Hedge Control Platform",
    "version": "0.3.0-phase3-step2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/cashflow/analytic": {
      "get": {
        "operationId": "get_cashflow_analytic_cashflow_analytic_get",
        "parameters": [
          {
            "in": "query",
            "name": "as_of_date",
            "required": true,
            "schema": {
              "format": "date",
              "title": "As Of Date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashFlowAnalyticResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Cashflow Analytic",
        "tags": [
          "CashFlow"
        ]
      }
    },
    "/cashflow/baseline/snapshots": {
      "get": {
        "operationId": "get_baseline_snapshot_cashflow_baseline_snapshots_get",
        "parameters": [
          {
            "in": "query",
            "name": "as_of_date",
            "required": true,
            "schema": {
              "format": "date",
              "title": "As Of Date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashFlowBaselineSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Baseline Snapshot",
        "tags": [
          "CashFlow"
        ]
      },
      "post": {
        "operationId": "create_baseline_snapshot_cashflow_baseline_snapshots_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashFlowBaselineSnapshotCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashFlowBaselineSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Baseline Snapshot",
        "tags": [
          "CashFlow"
        ]
      }
    },
    "/cashflow/contracts/{contract_id}/settle": {
      "post": {
        "operationId": "settle_hedge_contract_cashflow_contracts__contract_id__settle_post",
        "parameters": [
          {
            "in": "path",
            "name": "contract_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Contract Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HedgeContractSettlementCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HedgeContractSettlementResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Settle Hedge Contract",
        "tags": [
          "CashFlowLedger"
        ]
      }
    },
    "/cashflow/ledger": {
      "get": {
        "operationId": "list_ledger_entries_by_event_cashflow_ledger_get",
        "parameters": [
          {
            "in": "query",
            "name": "source_event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Event Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source_event_type",
            "required": false,
            "schema": {
              "default": "HEDGE_CONTRACT_SETTLED",
              "title": "Source Event Type",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CashFlowLedgerEntryRead"
                  },
                  "title": "Response List Ledger Entries By Event Cashflow Ledger Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Ledger Entries By Event",
        "tags": [
          "CashFlowLedger"
        ]
      }
    },
    "/cashflow/ledger/hedge-contracts/{contract_id}": {
      "get": {
        "operationId": "list_ledger_entries_for_contract_cashflow_ledger_hedge_contracts__contract_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "contract_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Contract Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CashFlowLedgerEntryRead"
                  },
                  "title": "Response List Ledger Entries For Contract Cashflow Ledger Hedge Contracts  Contract Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Ledger Entries For Contract",
        "tags": [
          "CashFlowLedger"
        ]
      }
    },
    "/cashflows": {
      "get": {
        "operationId": "list_cashflows_cashflows_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CashFlowRead"
                  },
                  "title": "Response List Cashflows Cashflows Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Cashflows",
        "tags": [
          "CashFlows"
        ]
      },
      "post": {
        "operationId": "create_cashflow_cashflows_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashFlowCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashFlowRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Cashflow",
        "tags": [
          "CashFlows"
        ]
      }
    },
    "/cashflows/{cashflow_id}": {
      "get": {
        "operationId": "get_cashflow_cashflows__cashflow_id__get",
        "parameters": [
          {
            "in": "query",
            "name": "_",
            "required": true,
            "schema": {
              "title": " ",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashFlowRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Cashflow",
        "tags": [
          "CashFlows"
        ]
      }
    },
    "/contracts/hedge": {
      "post": {
        "operationId": "create_hedge_contract_contracts_hedge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HedgeContractCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HedgeContractRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Hedge Contract",
        "tags": [
          "Contracts"
        ]
      }
    },
    "/contracts/hedge/{contract_id}": {
      "get": {
        "operationId": "get_hedge_contract_contracts_hedge__contract_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "contract_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Contract Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HedgeContractRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Hedge Contract",
        "tags": [
          "Contracts"
        ]
      }
    },
...
  }
}